Rename dist to distributed
