+++ mlx/distributed/ops.h
@@ -2,10 +2,10 @@
 
 #pragma once
 
-#include "mlx/dist/primitives.h"
+#include "mlx/distributed/primitives.h"
 
-namespace mlx::core::dist {
+namespace mlx::core::distributed {
 
 array all_reduce_sum(const array& x, std::shared_ptr<Group> group = nullptr);
 
-} // namespace mlx::core::dist
+} // namespace mlx::core::distributed
