+++ python/src/distributed.cpp
@@ -3,45 +3,46 @@
 #include <nanobind/nanobind.h>
 #include <nanobind/stl/optional.h>
 
-#include "mlx/dist/dist.h"
-#include "mlx/dist/ops.h"
+#include "mlx/distributed/distributed.h"
+#include "mlx/distributed/ops.h"
 
 namespace nb = nanobind;
 using namespace nb::literals;
 using namespace mlx::core;
 
-void init_dist(nb::module_& parent_module) {
+void init_distributed(nb::module_& parent_module) {
   auto m = parent_module.def_submodule(
-      "dist", "mlx.core.dist: Communication operations");
+      "distributed", "mlx.core.distributed: Communication operations");
 
-  nb::class_<dist::Group>(
+  nb::class_<distributed::Group>(
       m,
       "Group",
       R"pbcopy(
-        An :class:`mlx.core.dist.Group` represents a group of independent mlx
+        An :class:`mlx.core.distributed.Group` represents a group of independent mlx
         processes that can communicate.
       )pbcopy")
-      .def("rank", &dist::Group::rank, "Get the rank of this process")
-      .def("size", &dist::Group::size, "Get the size of the group");
+      .def("rank", &distributed::Group::rank, "Get the rank of this process")
+      .def("size", &distributed::Group::size, "Get the size of the group");
 
   m.def(
       "is_available",
-      &dist::is_available,
+      &distributed::is_available,
       R"pbdoc(
       Check if a communication backend is available.
       )pbdoc");
 
   m.def(
       "init",
-      &dist::init,
+      &distributed::init,
       R"pbdoc(
         Initialize the communication backend and create the global communication group.
       )pbdoc");
 
   m.def(
       "all_reduce_sum",
-      [](const array& x, std::optional<std::shared_ptr<dist::Group>> group) {
-        return dist::all_reduce_sum(x, group.value_or(nullptr));
+      [](const array& x,
+         std::optional<std::shared_ptr<distributed::Group>> group) {
+        return distributed::all_reduce_sum(x, group.value_or(nullptr));
       },
       "x"_a,
       nb::kw_only(),
